refactor: include replace_in_file in common crate
