+++ crates/pop-common/Cargo.toml
@@ -21,3 +21,4 @@ url.workspace = true
 
 [dev-dependencies]
 mockito.workspace = true
+tempfile.workspace = true
