+++ crates/pop-contracts/src/utils/helpers.rs
@@ -5,9 +5,6 @@ use contract_build::ManifestPath;
 use contract_extrinsics::BalanceVariant;
 use ink_env::{DefaultEnvironment, Environment};
 use std::{
-	collections::HashMap,
-	fs,
-	io::{Read, Write},
 	path::{Path, PathBuf},
 	str::FromStr,
 };
@@ -49,29 +46,6 @@ pub fn canonicalized_path(target: &Path) -> Result<PathBuf, Error> {
 		.map_err(|e| Error::IO(e))
 }
 
-/// Replaces occurrences of specified strings in a file with new values.
-///
-/// # Arguments
-///
-/// * `file_path` - A `PathBuf` specifying the path to the file to be modified.
-/// * `replacements` - A `HashMap` where each key-value pair represents
-///   a target string and its corresponding replacement string.
-///
-pub fn replace_in_file(file_path: PathBuf, replacements: HashMap<&str, &str>) -> Result<(), Error> {
-	// Read the file content
-	let mut file_content = String::new();
-	fs::File::open(&file_path)?.read_to_string(&mut file_content)?;
-	// Perform the replacements
-	let mut modified_content = file_content;
-	for (target, replacement) in &replacements {
-		modified_content = modified_content.replace(target, replacement);
-	}
-	// Write the modified content back to the file
-	let mut file = fs::File::create(&file_path)?;
-	file.write_all(modified_content.as_bytes())?;
-	Ok(())
-}
-
 #[cfg(test)]
 mod tests {
 	use super::*;
@@ -105,21 +79,6 @@ mod tests {
 		assert!(error_folder.is_err());
 		// Success case
 		canonicalized_path(temp_dir.path())?;
-		Ok(())
-	}
-
-	#[test]
-	fn test_replace_in_file() -> Result<(), Error> {
-		let temp_dir = tempfile::tempdir()?;
-		let file_path = temp_dir.path().join("file.toml");
-		let mut file = fs::File::create(temp_dir.path().join("file.toml"))?;
-		writeln!(file, "name = test, version = 5.0.0")?;
-		let mut replacements_in_cargo = HashMap::new();
-		replacements_in_cargo.insert("test", "changed_name");
-		replacements_in_cargo.insert("5.0.0", "5.0.1");
-		replace_in_file(file_path.clone(), replacements_in_cargo)?;
-		let content = fs::read_to_string(file_path).expect("Could not read file");
-		assert_eq!(content.trim(), "name = changed_name, version = 5.0.1");
 		Ok(())
 	}
 }
